+++ src/components/nodes/ClusterNodes.vue
@@ -62,9 +62,16 @@
   import { useElasticsearchRequest } from '../../composables/CallElasticsearch'
   import { useTranslation } from '../../composables/i18n'
   import { EsNode } from '../../types/types.ts'
+  import { useConnectionStore } from '../../store/connection.ts'
 
   const t = useTranslation()
-  const CAT_METHOD_PARAMS = {
+  const connectionStore = useConnectionStore()
+
+  type methodParams = {
+    h: string[],
+    full_id?: boolean
+  }
+  const CAT_METHOD_PARAMS: methodParams = {
     h: [
       'ip',
       'id',
@@ -84,8 +91,11 @@
       'disk.used_percent',
       'disk.used',
       'disk.total'
-    ],
-    full_id: true
+    ]
+  }
+
+  if (connectionStore.activeCluster && parseInt(connectionStore.activeCluster.majorVersion) > 5) {
+    CAT_METHOD_PARAMS.full_id = true
   }
 
   const { load, requestState, data } = useElasticsearchRequest<EsNode[]>('catNodes', CAT_METHOD_PARAMS)