+++ tests/e2e/helpers.ts
@@ -1,4 +1,6 @@
-export const setupClusterConnection = async page => {
+import { Page } from '@playwright/test'
+
+export const setupClusterConnection = async (page: Page) => {
   await page.addInitScript(() => {
     window.localStorage.clear()
 
@@ -25,9 +27,9 @@ export const setupClusterConnection = async page => {
     window.localStorage.setItem('theme', '{"dark":true}')
   })
 
-  await page.goto('http://localhost:5173')
+  await page.goto('http://localhost:5175')
 }
 
 export const openElasticvue = page => {
-  return page.goto('http://localhost:5173/')
+  return page.goto('http://localhost:5175/')
 }