+++ tests/e2e/mocks/elastic_7/index.ts
@@ -1,0 +1,8 @@
+import { Page } from '@playwright/test'
+import { mockElasticHome } from './home'
+import { mockElasticNodes } from './nodes'
+
+export const mockElastic7 = async (page: Page, { health }: { health: string } = { health: 'green' }) => {
+  await mockElasticHome(page, { health })
+  await mockElasticNodes(page)
+}