+++ tests/e2e/tests/pages/ClusterManagement.spec.ts
@@ -1,48 +1,58 @@
-import { test, expect } from '@playwright/test'
+import { test, expect, Page } from '@playwright/test'
 import { setupClusterConnection } from '../../helpers'
-import { mockElastic } from '../../mocks/api'
+import { withElastic } from '../../mocks'
 
-const setupClusterSelectionTest = async page => {
-  await mockElastic(page)
+const setupClusterSelectionTest = async (page: Page) => {
   await setupClusterConnection(page)
   await page.getByTestId('cluster-selection').click()
-  return await page.getByTestId('cluster-table')
+  return page.getByTestId('cluster-table')
 }
 
-test.describe('ClusterManagement', () => {
-  test('table contains cluster names', async ({ page }) => {
-    const table = await setupClusterSelectionTest(page)
-    const rows = await table.locator('tbody').locator('tr')
+withElastic(({ mockElastic, elastic }) => {
+  test.describe(`elasticsearch ${elastic.version}`, () => {
+    test.describe('ClusterManagement', () => {
+      test('table contains cluster names', async ({ page }) => {
+        const table = await setupClusterSelectionTest(page)
 
-    await expect(rows).toHaveCount(2)
-    await expect(table).toContainText('default cluster')
-    await expect(table).toContainText('dev cluster')
-  })
+        await mockElastic(page)
+        const rows = table.locator('tbody').locator('tr')
 
-  test('can filter cluster names', async ({ page }) => {
-    const table = await setupClusterSelectionTest(page)
-    await expect(table).toContainText('dev cluster')
-    await page.getByTestId('cluster-table-filter').fill('default cluster')
+        await expect(rows).toHaveCount(2)
+        await expect(table).toContainText('default cluster')
+        await expect(table).toContainText('dev cluster')
+      })
 
-    await expect(table).toContainText('default cluster')
-    await expect(table).not.toContainText('dev cluster')
-  })
+      test('can filter cluster names', async ({ page }) => {
+        await mockElastic(page)
+        const table = await setupClusterSelectionTest(page)
 
-  test('can change cluster', async ({ page }) => {
-    await setupClusterSelectionTest(page)
-    await page.getByTestId('cluster-table-row-1').click()
+        await expect(table).toContainText('dev cluster')
+        await page.getByTestId('cluster-table-filter').fill('default cluster')
 
-    expect(page.url()).toContain('/cluster/1')
-  })
+        await expect(table).toContainText('default cluster')
+        await expect(table).not.toContainText('dev cluster')
+      })
 
-  test('can rename cluster', async ({ page }) => {
-    const table = await setupClusterSelectionTest(page)
-    await page.getByTestId('cluster-table-row-1').getByTestId('cluster-edit').click()
+      test('can change cluster', async ({ page }) => {
+        await mockElastic(page)
+        await setupClusterSelectionTest(page)
+        await page.getByTestId('cluster-table-row-1').click()
 
-    const newName = 'foo'
-    await page.getByTestId('cluster-edit-name').fill(newName)
-    await page.getByTestId('cluster-edit-save').click()
+        expect(page.url()).toContain('/cluster/1')
+      })
 
-    await expect(table).toContainText(newName)
+      test('can rename cluster', async ({ page }) => {
+        const table = await setupClusterSelectionTest(page)
+
+        await mockElastic(page)
+        await page.getByTestId('cluster-table-row-1').getByTestId('cluster-edit').click()
+
+        const newName = 'foo'
+        await page.getByTestId('cluster-edit-name').fill(newName)
+        await page.getByTestId('cluster-edit-save').click()
+
+        await expect(table).toContainText(newName)
+      })
+    })
   })
 })